move logUncaught out of HelpfulExecutors
